pyproject.toml
[tool.poetry]
name = "minutes-tui"
version = "0.1.0"
description = ""
authors = []
packages = [
    { include = "minutes", from="src" },
]

[tool.poetry.dependencies]
python = "^3.7"
textual = "^0.16.0"
pyxdg = "^0.28"
sqlmodel = "^0.0.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[tool.poetry.scripts]
minutes = "minutes.minutes:run"


[tool.poetry.group.dev.dependencies]
textual = {extras = ["dev"], version = "^0.16.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
